--- a/src/Http/perf/Microbenchmarks/AssemblyInfo.cs
+++ b/src/Http/perf/Microbenchmarks/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+[assembly: BenchmarkDotNet.Attributes.AspNetCoreBenchmark]