--- a/src/Http/perf/Microbenchmarks/Microsoft.AspNetCore.Http.Abstractions.Microbenchmarks.csproj
+++ b/src/Http/perf/Microbenchmarks/Microsoft.AspNetCore.Http.Abstractions.Microbenchmarks.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Reference Include="BenchmarkDotNet" />
+    <Reference Include="Microsoft.AspNetCore.BenchmarkRunner.Sources" />
+    <Reference Include="Microsoft.AspNetCore.Http.Abstractions" />
+    <Reference Include="Microsoft.AspNetCore.Http" />
+  </ItemGroup>
+</Project>