Method |     Mean |     Error |    StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
-------------------------------- |---------:|----------:|----------:|------------:|-------:|---------:|-------:|----------:|
SplitSingleHeaderOriginal | 344.9 ns |  6.914 ns | 18.455 ns | 2,899,279.9 |   1.00 |     0.00 | 0.0057 |     192 B |
SplitSingleHeader | 192.0 ns |  4.087 ns |  4.542 ns | 5,207,375.6 |   0.56 |     0.03 | 0.0038 |     120 B |

Method |     Mean |     Error |    StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
-------------------------------- |---------:|----------:|----------:|------------:|-------:|---------:|-------:|----------:|
SplitSingleQuotedHeaderOriginal | 364.5 ns |  6.933 ns |  6.485 ns | 2,743,668.9 |   1.00 |     0.00 | 0.0072 |     232 B |
SplitSingleQuotedHeader | 221.6 ns |  4.455 ns |  6.245 ns | 4,513,130.2 |   0.61 |     0.02 | 0.0048 |     160 B |

Method |     Mean |     Error |    StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
-------------------------------- |---------:|----------:|----------:|------------:|-------:|---------:|-------:|----------:|
SplitDoubleHeaderOriginal | 441.8 ns |  8.725 ns | 10.715 ns | 2,263,329.0 |   1.00 |     0.00 | 0.0072 |     200 B |
SplitDoubleHeader | 259.9 ns |  5.057 ns |  7.253 ns | 3,847,705.3 |   0.59 |     0.02 | 0.0038 |     128 B |

Method |     Mean |     Error |    StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
-------------------------------- |---------:|----------:|----------:|------------:|-------:|---------:|-------:|----------:|
SplitManyHeadersOriginal | 852.8 ns | 11.214 ns |  9.364 ns | 1,172,649.9 |   1.00 |     0.00 | 0.0105 |     320 B |
SplitManyHeaders | 615.8 ns | 11.767 ns | 14.451 ns | 1,623,843.3 |   0.72 |     0.02 | 0.0076 |     248 B |

Also did an experimental change and got:

Method |     Mean |     Error |    StdDev |   Median |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
------------------------------------|---------:|----------:|----------:|---------:|------------:|-------:|---------:|-------:|----------:|
SplitSingleHeaderOriginal | 388.1 ns | 16.972 ns | 50.043 ns | 374.8 ns | 2,576,895.9 |   1.00 |     0.00 | 0.0057 |     192 B |
SplitSingleHeader | 191.9 ns |  3.781 ns |  4.045 ns | 190.6 ns | 5,212,160.9 |   0.50 |     0.06 | 0.0038 |     120 B |
SplitSingleHeaderExperimental | 142.4 ns |  2.996 ns |  6.188 ns | 141.8 ns | 7,023,061.2 |   0.37 |     0.05 | 0.0010 |      32 B |

But it relied on moving some of the parsing logic into the `HeaderSplit` method.