@davidfowl 2019-04-12
This PR:

```
                  Method |     Mean |     Error |    StdDev |        Op/s |  Gen 0 | Allocated |
------------------------ |---------:|----------:|----------:|------------:|-------:|----------:|
       SplitSingleHeader | 154.3 ns |  3.957 ns | 11.225 ns | 6,480,460.2 | 0.0038 |     120 B |
 SplitSingleQuotedHeader | 173.1 ns |  3.315 ns |  3.101 ns | 5,776,599.5 | 0.0052 |     160 B |
       SplitDoubleHeader | 226.7 ns |  5.713 ns | 16.665 ns | 4,410,972.3 | 0.0043 |     128 B |
        SplitManyHeaders | 615.6 ns | 19.972 ns | 58.574 ns | 1,624,456.3 | 0.0067 |     248 B |
```    

My slight modification:

```     
                  Method |     Mean |     Error |    StdDev |    Median |        Op/s |  Gen 0 | Allocated |
------------------------ |---------:|----------:|----------:|----------:|------------:|-------:|----------:|
       SplitSingleHeader | 101.4 ns |  2.055 ns |  5.695 ns |  99.54 ns | 9,866,116.7 |      - |       0 B |
 SplitSingleQuotedHeader | 146.0 ns |  4.157 ns | 11.993 ns | 144.32 ns | 6,849,742.2 | 0.0014 |      40 B |
       SplitDoubleHeader | 204.0 ns |  4.825 ns | 14.226 ns | 202.31 ns | 4,901,978.0 | 0.0010 |      40 B |
        SplitManyHeaders | 707.5 ns | 14.122 ns | 18.362 ns | 704.63 ns | 1,413,467.7 | 0.0086 |     280 B |
```

SplitManyHeaders is slower.