@anurse 2019-04-12
It looks like this PR will make https://github.com/aspnet/AspNetCore/issues/3737 irrelevant since you're inlining the implementation.